Split all classes into individual files
Split Index, Country, Legislature, and LegislativePeriod out into
their own files.
